requirements.txt
fastapi
uvicorn
openai
faiss-cpu
numpy
pydantic
pypdf
